refactor : dealerProfit Long으로 변경

+++ src/main/java/domain/BlackjackResult.java
@@ -9,33 +9,33 @@ public class BlackjackResult {
 
     private static final int DEALER_PROFIT_INVERSION = -1;
 
-    private final Map<String, Integer> gamblerProfits;
+    private final Map<String, Long> gamblerProfits;
     private final long dealerProfit;
 
-    private BlackjackResult(Map<String, Integer> gamblerProfits, long dealerProfit) {
+    private BlackjackResult(Map<String, Long> gamblerProfits, long dealerProfit) {
         this.gamblerProfits = gamblerProfits;
         this.dealerProfit = dealerProfit;
     }
 
     public static BlackjackResult from(Dealer dealer, Gamblers gamblers) {
-        Map<String, Integer> profits = calculateGamblerProfits(dealer, gamblers);
+        Map<String, Long> profits = calculateGamblerProfits(dealer, gamblers);
         long dealerProfit = calculateDealerProfit(profits);
         return new BlackjackResult(profits, dealerProfit);
     }
 
-    private static Map<String, Integer> calculateGamblerProfits(Dealer dealer, Gamblers gamblers) {
-        Map<String, Integer> profits = new LinkedHashMap<>();
+    private static Map<String, Long> calculateGamblerProfits(Dealer dealer, Gamblers gamblers) {
+        Map<String, Long> profits = new LinkedHashMap<>();
 
         gamblers.forEach(gambler -> {
             MatchResult matchResult = MatchResult.of(gambler, dealer);
-            int reward = gambler.calculateReward(matchResult);
+            long reward = gambler.calculateReward(matchResult);
             profits.put(gambler.getName(), reward);
         });
 
         return profits;
     }
 
-    private static long calculateDealerProfit(Map<String, Integer> gamblerProfits) {
+    private static long calculateDealerProfit(Map<String, Long> gamblerProfits) {
         long totalGamblerProfit = gamblerProfits.values().stream()
                 .mapToLong(Long::valueOf)
                 .sum();
@@ -43,7 +43,7 @@ public class BlackjackResult {
         return totalGamblerProfit * DEALER_PROFIT_INVERSION;
     }
 
-    public Map<String, Integer> gamblerProfits() {
+    public Map<String, Long> gamblerProfits() {
         return Map.copyOf(gamblerProfits);
     }
